Cleanup: Mikktspace: use sqrtf() function with float type
Pull Request: https://projects.blender.org/blender/blender/pulls/121703

+++ intern/mikktspace/mikk_float3.hh
@@ -64,7 +64,7 @@ struct float3 {
 
   float length() const
   {
-    return sqrt(length_squared());
+    return sqrtf(length_squared());
   }
 
   static float distance(const float3 &a, const float3 &b)
